use fast types

+++ src/cmd/bench-swiper/main.cpp
@@ -18,7 +18,7 @@ using namespace std::string_view_literals;
 
 #include "swiper/swiper.hpp"
 
-constexpr auto trials = 1 << 30;
+constexpr auto trials = int_fast32_t(1 << 30);
 
 int main() {
     #if defined(_WIN32)
